+++ src/SunburstSettings.ts
@@ -41,7 +41,8 @@ export const enum SunburstObjectNames {
     LegendTitle = "legendTitle",
     Group = "group",
     Color = "colorsGroup",
-    Label = "labelsGroup"
+    Label = "labelsGroup",
+    Value = "valuesGroup"
 }
 
 class BaseFontCardSettings extends formattingSettings.FontControl {
@@ -156,7 +157,7 @@ class SunburstCentralLabelSettings extends FormattingSettingsCompositeCard {
     public displayNameKey: string = "Visual_CentralLabel";
 }
 
-class LabelsGroup extends FormattingSettingsCard {
+export class LabelsGroup extends FormattingSettingsCard {
     public defaultShowDataLabels: boolean = true;
     public defaultLabelFontSize: number = 12;
     public defaultLabelColor: string = "#000000";
@@ -181,6 +182,30 @@ class LabelsGroup extends FormattingSettingsCard {
     slices: FormattingSettingsSlice[] = [this.font, this.labelColor];
 }
 
+class ValuesGroup extends FormattingSettingsCard {
+    public defaultFormat: string = ": #,#.#";
+    public defaultShowDataValues: boolean = false;
+
+    public showDataValues = new formattingSettings.ToggleSwitch({
+        name: "showDataValues",
+        displayNameKey: "Visual_ShowDataValues",
+        value: this.defaultShowDataValues,
+    });
+
+    public format: formattingSettings.TextInput = new formattingSettings.TextInput({
+        name: "format",
+        displayNameKey: "Visual_Format",
+        descriptionKey: "Visual_FormatDescription",
+        value: this.defaultFormat,
+        placeholder: ""
+    });
+
+    topLevelSlice: formattingSettings.ToggleSwitch = this.showDataValues;
+    name: string = SunburstObjectNames.Value;
+    displayNameKey: string = "Visual_ShowDataValues";
+    slices: FormattingSettingsSlice[] = [this.format];
+}
+
 class ColorsGroup extends FormattingSettingsCard {
     name: string = SunburstObjectNames.Color;
     displayNameKey: string = "Visual_Colors";
@@ -189,9 +214,10 @@ class ColorsGroup extends FormattingSettingsCard {
 
 class SunburstGroupSettings extends FormattingSettingsCompositeCard {
     public labels = new LabelsGroup();
+    public values = new ValuesGroup();
     public colors = new ColorsGroup();
 
-    public groups: FormattingSettingsGroup[] = [this.labels, this.colors];
+    public groups: FormattingSettingsGroup[] = [this.labels, this.values, this.colors];
     public name: string = SunburstObjectNames.Group ;
     public displayNameKey: string = "Visual_Groups";
     public analyticsPane: boolean = false;
@@ -318,6 +344,9 @@ export class SunburstSettings extends FormattingSettingsModel {
 
     public cards: Array<FormattingSettingsCard> = [this.centralLabel, this.group, this.tooltip, this.legend];
 
+    public parseSettings(): void {
+        this.group.values.visible = this.group.labels.showDataLabels.value;
+    }
     public setSlicesForTopCategoryColorPickers(topCategories: SunburstDataPoint[], LegendPropertyIdentifier: powerbiVisualsApi.DataViewObjectPropertyIdentifier, ColorHelper) {
         if (topCategories && topCategories.length > 0) {
             topCategories.forEach((category: SunburstDataPoint) => {